requirements.txt
azure-functions
azf-wsgi==0.3.1
flask==1.1.2
